Quick first version.
I copied a lot of the code from GrassyKnight, because there's several
problems that are the same between the two projects. Namely:

1) Uniquely identifying game objects (helped by the *Key class).
2) Associating some state with game objects (helped by the *DB class).
3) Storing that state _exactly_ in save files (helped by the
   serialization code).

So I was able to get this version of the code doing something useful
fairly quickly. As it stands, the mod will prevent enemies from
dropping geo after they die once. However, the mod isn't yet aware of
spawned enemies.