Another WIP: sketched out most of it.
I'm not sure I'll finish this, it's pretty straightforward at this
point... but straightforward is fairly boring which isn't waht I'm
looking for from my side projects.

But if I do come back to this. I haven't tried to compile most of this
code so first step will be dealing with all the compile-time errors.